Staging in preparation for releasing Aerie 0.4.0

+++ merlin-sdk/src/main/java/gov/nasa/jpl/ammos/mpsa/aerie/merlinsdk/typemappers/DoubleParameterMapper.java
@@ -7,7 +7,7 @@ import gov.nasa.jpl.ammos.mpsa.aerie.merlinsdk.utilities.Result;
 public final class DoubleParameterMapper implements ParameterMapper<Double> {
   @Override
   public ParameterSchema getParameterSchema() {
-    return ParameterSchema.DOUBLE;
+    return ParameterSchema.REAL;
   }
 
   @Override
